+++ Cargo.toml
@@ -4,5 +4,6 @@ members = [
   "clearned_example",
   "clearned_core",
   "clearned_bench",
+  "clearned_derive",
   "clearned"
 ]
