+++ clearned/Cargo.toml
@@ -1,13 +1,8 @@
 [package]
 name = "clearned"
-description = "Proc macro for building clearned data structures."
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-proc-macro = true
-
 [dependencies]
-syn = { version = "2.0", features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
+clearned_derive = { path = "../clearned_derive" }
+clearned_core = { path = "../clearned_core" }
