+++ clearned_example/Cargo.toml
@@ -5,4 +5,3 @@ edition = "2021"
 
 [dependencies]
 clearned = { path = "../clearned" }
-clearned_core = { path = "../clearned_core" }
